requirements.txt
python-fasthtml
markdown
python-frontmatter
pyyaml
pytz
jinja2
